test: Introduce "live" acceptance test for tpf
